refactor: move HitRecord to its own module
