+++ clovers/src/hitable.rs
@@ -8,46 +8,14 @@ use crate::objects::{GLTFTriangle, GLTF};
 use crate::{
     aabb::AABB,
     bvh::BVHNode,
-    materials::MaterialTrait,
     objects::{Boxy, ConstantMedium, MovingSphere, Quad, RotateY, Sphere, Translate, Triangle},
     ray::Ray,
     wavelength::Wavelength,
-    Direction, Displacement, Float, Position,
+    Direction, Displacement, Float, HitRecord, Position,
 };
 
 use enum_dispatch::enum_dispatch;
 use rand::rngs::SmallRng;
-
-/// Represents a ray-object intersection, with plenty of data about the intersection.
-#[derive(Clone, Debug)]
-pub struct HitRecord<'a> {
-    /// Distance from the ray origin to the hitpoint
-    pub distance: Float,
-    /// 3D coordinate of the hitpoint
-    pub position: Position,
-    /// Surface normal from the hitpoint
-    pub normal: Direction,
-    /// U surface coordinate of the hitpoint
-    pub u: Float,
-    /// V surface coordinate of the hitpoint
-    pub v: Float,
-    /// Reference to the material at the hitpoint
-    pub material: &'a dyn MaterialTrait,
-    /// Is the hitpoint at the front of the surface
-    pub front_face: bool,
-}
-
-impl<'a> HitRecord<'a> {
-    /// Helper function for getting normals pointing at the correct direction. TODO: consider removal?
-    pub fn set_face_normal(&mut self, ray: &Ray, outward_normal: Direction) {
-        self.front_face = ray.direction.dot(&outward_normal) < 0.0;
-        if self.front_face {
-            self.normal = outward_normal;
-        } else {
-            self.normal = -outward_normal;
-        }
-    }
-}
 
 /// Enumeration of all runtime entities that can be intersected aka "hit" by a [Ray].
 #[enum_dispatch(HitableTrait)]
