+++ clovers/src/objects/rotate.rs
@@ -2,10 +2,10 @@
 
 use crate::{
     aabb::AABB,
-    hitable::{HitRecord, Hitable, HitableTrait},
+    hitable::{Hitable, HitableTrait},
     ray::Ray,
     wavelength::Wavelength,
-    Box, Direction, Float, Position, Vec3,
+    Box, Direction, Float, HitRecord, Position, Vec3,
 };
 use nalgebra::Unit;
 use rand::rngs::SmallRng;
